fix(account): when display public keys from wallet prefix index and alias (#2020)
* fix(account): when display public keys from wallet prefix index and alias

* fix(account): show alias only any key has alias

* refactor: move convert keys to choices logic to separate helper

* chore: remove idx prefix from key index

* refactor: use tableFormatPrompt helper to display table format selection for keys
