+++ packages/tools/kadena-cli/src/account/commands/accountAddFromWallet.ts
@@ -1,55 +1,11 @@
-import { Option } from 'commander';
-import { z } from 'zod';
-
-import { MULTI_SELECT_INSTRUCTIONS } from '../../constants/global.js';
-import type { IWallet } from '../../services/wallet/wallet.types.js';
 import { assertCommandError } from '../../utils/command.util.js';
 import { createCommand } from '../../utils/createCommand.js';
-import { createOption } from '../../utils/createOption.js';
 import { globalOptions } from '../../utils/globalOptions.js';
 import { log } from '../../utils/logger.js';
-import { checkbox } from '../../utils/prompts.js';
 import { accountOptions } from '../accountOptions.js';
 import { addAccount } from '../utils/addAccount.js';
-import { displayAddAccountSuccess, isEmpty } from '../utils/addHelpers.js';
+import { displayAddAccountSuccess } from '../utils/addHelpers.js';
 import { validateAndRetrieveAccountDetails } from '../utils/validateAndRetrieveAccountDetails.js';
-
-const selectPublicKeys = createOption({
-  key: 'publicKeys' as const,
-  defaultIsOptional: false,
-  async prompt(args) {
-    const wallet = args.walletNameConfig as IWallet;
-    const publicKeysList = wallet.keys.reduce(
-      (acc, key) => acc.concat([key.publicKey]),
-      [] as string[],
-    );
-    const selectedKeys = await checkbox({
-      message: 'Select public keys to add to account:',
-      choices: publicKeysList.map((key) => ({ value: key })),
-      pageSize: 10,
-      instructions: MULTI_SELECT_INSTRUCTIONS,
-      validate: (input) => {
-        if (input.length === 0) {
-          return 'Please select at least one public key';
-        }
-
-        return true;
-      },
-    });
-    return selectedKeys.join(',');
-  },
-  expand: async (publicKeys: string): Promise<string[]> => {
-    const keys = publicKeys.split(',');
-    return keys
-      .map((key: string) => key.trim())
-      .filter((key: string) => !isEmpty(key));
-  },
-  validation: z.string(),
-  option: new Option(
-    '-k, --public-keys <publicKeys>',
-    'Public keys to add to account',
-  ),
-});
 
 export const createAddAccountFromWalletCommand = createCommand(
   'add-from-wallet',
@@ -60,7 +16,7 @@ export const createAddAccountFromWalletCommand = createCommand(
     accountOptions.fungible(),
     globalOptions.networkSelect(),
     globalOptions.chainId(),
-    selectPublicKeys(),
+    accountOptions.selectPublicKeys(),
     accountOptions.predicate(),
     accountOptions.accountOverwrite(),
   ],
